2015: d22p1 d22p2

+++ 2015/src/day21/day21.h
@@ -3,5 +3,5 @@
 #include <stdint.h>
 
 void day21_main(void);
-uint16_t day21_part1(const char* const*);
-uint16_t day21_part2(const char* const*);
+uint8_t day21_part1(const char* const*);
+uint8_t day21_part2(const char* const*);
